Update
Added :
- Health system.
- MainMenu begin, by the way use blueprint to start the local server
I'll manage C++...
- Added Super's HUD, and fixed some strange things with DrawSimpleText
also redirected health/max health vars to the local UModCharacter.
- New GameMode (for future lua integration) function :
OnPlayerDeath(AUModCharacter *player).
- Prepared for loading systems shemes (will use virtual directory system
to manage everything, with a max of 32768 lua files)...

+++ Source/UMod/UModHUD.h
@@ -14,7 +14,7 @@ public:
 	/** Primary draw call for the HUD */
 	virtual void DrawHUD() override;
 
-	void DrawSimpleText(FString String, float X, float Y, float Size, FColor Color, float Alligment);
+	void DrawSimpleText(FString String, float X, float Y, float Size, FColor Color, uint8 Alligment);
 	void DrawRect(float X, float Y, float W, float H, FLinearColor Color);
 
 	FVector2D ScreenSize();
@@ -24,6 +24,5 @@ private:
 	class UTexture2D* CrosshairTex;
 
 	class UFont* HUDFont;
-	
 };
